+++ cpp/dataset_stats.hpp
@@ -144,11 +144,9 @@ class DatasetStats : public Dataset {
       uint64_t num_chunks = 0, num_traces = reader_->GetStatsSize();
 
       // Sum metrics for avg and var
-      // avg = \sum x / n
-      // var = - avg^2 + \sum x^2 / n
-      uint64_t total_usage = 0, total_usage_sq = 0;
-      uint64_t total_lifetime = 0, total_lifetime_sq = 0;
-      uint64_t total_useful_lifetime = 0, total_useful_lifetime_sq = 0;
+      float total_usage = 0, total_usage_sq = 0;
+      float total_lifetime = 0, total_lifetime_sq = 0;
+      float total_useful_lifetime = 0, total_useful_lifetime_sq = 0;
 
       for (size_t i = 0; i < num_traces; ++i) {
         const TraceStat& t = Trace(i);
@@ -166,14 +164,16 @@ class DatasetStats : public Dataset {
       }
 
       // Compute avg
-      float usage_avg = total_usage / (float)num_traces;
-      float lifetime_avg = total_lifetime / (float)num_traces;
-      float useful_lifetime_avg = total_useful_lifetime / (float)num_traces;
+      // avg = \sum x / n
+      float usage_avg = total_usage / num_traces;
+      float lifetime_avg = total_lifetime / num_traces;
+      float useful_lifetime_avg = total_useful_lifetime / num_traces;
 
       // Comput var
-      float usage_var = total_usage_sq / (float)num_traces - usage_avg * usage_avg;
-      float lifetime_var = total_lifetime_sq / (float)num_traces - lifetime_avg * lifetime_avg;
-      float useful_lifetime_var = total_useful_lifetime_sq / (float)num_traces - useful_lifetime_avg * useful_lifetime_avg;
+      // var = - avg^2 + \sum x^2 / n
+      float usage_var = total_usage_sq / num_traces - usage_avg * usage_avg;
+      float lifetime_var = total_lifetime_sq / num_traces - lifetime_avg * lifetime_avg;
+      float useful_lifetime_var = total_useful_lifetime_sq / num_traces - useful_lifetime_avg * useful_lifetime_avg;
 
       uint64_t total_time = max_time_ - min_time_;
